--- todo/form.py
@@ -1,8 +1,0 @@
-from todo.models import Todo
-from django import forms
-
-
-class TodoForm(forms.ModelForm):
-    class Meta:
-        model = Todo
-        fields = ['title', 'description', 'important']
