--- todo/models.py
@@ -1,11 +1,0 @@
-from django.db import models
-
-class Todo(models.Model):
-    title = models.CharField(max_length=100)
-    description = models.TextField(blank=True)
-    created = models.DateTimeField(auto_now_add=True)
-    completed = models.BooleanField(default=False)
-    important = models.BooleanField(default=False)
-
-    def __str__(self):
-        return self.title
