+++ todoproject/todo/form.py
@@ -1,0 +1,8 @@
+from todo.models import Todo
+from django import forms
+
+
+class TodoForm(forms.ModelForm):
+    class Meta:
+        model = Todo
+        fields = ['title', 'description', 'important']
